py.requirements.txt
# ============================================================================
# PYTHON PACKAGE REQUIREMENTS FOR: explore.guardrails
# ============================================================================
# MAYBE: vale -- https://vale.sh/docs/topics/config/

guardrails-ai >= 0.5.4
cucumber-expressions
rstr

# pydantic >= 2.7.3

# -- BUILD-SYSTEM
# PREPARED: invoke
# PREPARED: git+https://github.com/jenisys/invoke-cleanup@v0.3.7

# -- DEVELOP:
ruff
pyright
codespell >= 2.3.0

# -- TESTING: Unit tests and behave self-tests.
pytest >= 5.0
pytest-html >= 2.0
